refactor: separate sidebar from chat
