+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev --turbopack",
+    "dev": "rimraf .next && next dev --turbopack",
     "build": "next build",
     "start": "next start",
     "lint": "next lint"
