+++ src/app/_components/index.ts
@@ -7,3 +7,7 @@ export * from "./SetupForm";
 export * from "./ThemeChanger";
 export * from "./Usage";
 export * from "./Spinner";
+export * from "./Spinner";
+export * from "./Spinner";
+
+export * from "./hooks/localStorage.hook";
